third_party/blink/web_tests/fast/events/before-unload-return-value-from-listener-expected.txt

Tests the returnValue attribute of the BeforeUnloadEvent from an event listener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testEvent.__proto__ is BeforeUnloadEvent.prototype
PASS testEvent.returnValue is ""
PASS successfullyParsed is true

TEST COMPLETE
